homeworks-optional/actor_recognition/requirements.txt
numpy
scipy
insightface
pillow
filterpy
ffmpeg-python
numba
onnxruntime-gpu
click
